refactor(extension/podman): mv compatibility-mode.ts
